paused

+++ Chapter_02/web_fastapi/data/explorer.py
@@ -1,4 +1,4 @@
-from .init import curs
+from init import (conn, curs, IntegrityError)
 from model.explorer import Explorer
 
 curs.execute("""create table if not exists explorer(
@@ -7,16 +7,21 @@ curs.execute("""create table if not exists explorer(
             description text) """)
 
 def row_to_model(row: tuple) -> Explorer:
-    return Explorer(name=row[0], country=row[1], description=row[2])
+    name, country, description = row
+    return Explorer(name=name, country=country, description=description)
 
 def model_to_dict(explorer: Explorer) -> dict:
-    return explorer.dict() if explorer else None
+    return explorer.dict()
 
 def get_one(name: str) -> Explorer:
     qry = "select * from explorer where name = :name"
     params = {"name": name}
     curs.execute(qry, params)
-    return row_to_model(curs.fetchone())
+    row = curs.fetchone()
+    if row:
+        return row_to_model(row)
+    else:
+        raise Missing(msg=f"Explorer {name} not found")
 
 def get_all() -> list[Explorer]:
     qry = "select * from explorer"
@@ -24,21 +29,30 @@ def get_all() -> list[Explorer]:
     return [row_to_model(row) for row in curs.fetchall()]
 
 def create(explorer: Explorer) -> Explorer:
+    if not explorer: return None
     qry = """insert into explorer(name, country, description) values(:name, :country, :description)"""
     params = model_to_dict(explorer)
-    _ = curs.execute(qry, params)
+    try:
+        curs.execute(qry, params)
+    except IntegrityError:
+        raise Duplicate(msg=f"Explorer {explorer.name} already exists")
     return get_one(explorer.name)
 
 def modify(name: str, explorer: Explorer) -> Explorer:
+    if not (name and explorer): return None
     qry = """update explorer set country=:country, name=:name, description=:description where name=:name_orig"""
     params = model_to_dict(explorer)
     params["name_orig"] = explorer.name
-    _ = curs.execute(qry, params)
-    explorer2 = get_one(explorer.name)
-    return explorer2
+    curs.execute(qry, params)
+    if curs.rowcount == 1:
+        return get_one(explorer.name)
+    else:
+        raise Missing(msg=f"Explorer {name} not found")
 
-def delete(explorer: Explorer) -> bool:
+def delete(name: str):
+    if not name: return False
     qry = "delete from explorer where name = :name"
-    params = {"name": explorer.name}
-    res = curs.execute(qry, params)
-    return bool(res)
+    params = {"name": name}
+    curs.execute(qry, params)
+    if curs.rowcount != 1:
+        raise Missing(msg=f"Explorer {name} not found")
